Week 3 - Regression Analysis using HADS data
Analyzing different variables in the Housing Affordability Data System. The statistical significance of various variables were determined, taking logarithmic transformations of some variables to increase the correlation coefficient. This analysis enables us to determine which independent variables have the most impact in influencing the market value of the housing unit.